moved non-view components into components folder
