+++ src/views/Bots/BotsInfo/BotsInfo.js
@@ -5,8 +5,8 @@ import ErrorHOC from '../../../HOCs/ErrorHOC'
 import Web3HOC from '../../../HOCs/Web3HOC'
 import getDxService from '../../../services/dxService'
 
-import ErrorPre from '../../Error'
-import Loading from '../../Loading'
+import ErrorPre from '../../../components/Error'
+import Loading from '../../../components/Loading'
 
 import { from } from 'rxjs'
