+++ src/views/Markets/MarketList/MarketList.js
@@ -9,9 +9,9 @@ import { PageWrapper, PageFilter } from '../../../containers'
 import ErrorHOC from '../../../HOCs/ErrorHOC'
 import Web3HOC from '../../../HOCs/Web3HOC'
 
-import ErrorPre from '../../Error'
-import Loading from '../../Loading'
-import ColourKey from '../../ColourKey'
+import ErrorPre from '../../../components/Error'
+import Loading from '../../../components/Loading'
+import ColourKey from '../../../components/ColourKey'
 
 import getDxService from '../../../services/dxService'
