+++ src/views/Safes/index.js
@@ -7,8 +7,8 @@ import { PageFilter, PageWrapper } from '../../containers'
 import ErrorHOC from '../../HOCs/ErrorHOC'
 import Web3HOC from '../../HOCs/Web3HOC'
 
-import Loading from '../Loading'
-import ErrorPre from '../Error'
+import Loading from '../../components/Loading'
+import ErrorPre from '../../components/Error'
 
 import getDxService from '../../services/dxService'
